DedicatedServer
Adding build script and Target file for building the game as a dedicated server. Installed builds (rocket) do not ship with UE4Server dll's and you must run against an engine pulled from github source to compile a dedicated server. Script needs updating to point to source build directory for it to work.